feat: integrate GitHub OAuth for repository syncing
- Add GitHub OAuth service using Raycast built-in provider
- Create OAuth-wrapped repository management components:
  - manage-repos-oauth.tsx with OAuth authentication
  - repository-details-oauth.tsx with OAuth token access
- Update syncRepositoryFiles() to accept OAuth access token
- Replace manual GitHub token preference with OAuth flow
- Fix TypeScript errors and build issues:
  - Fix Icon.XMark to Icon.Xmark
  - Fix Icon.GitBranch to Icon.Code
  - Remove updatedAt property from UserRepository interface
  - Update repository-details.tsx to handle missing OAuth token
- Update package.json description to mention OAuth integration
- Maintain backward compatibility with existing functionality
- OAuth scope includes: 'repo read:org read:user' for repository access
- Users can now authenticate with GitHub seamlessly for repository syncing
